tests: factor common HostMemory / MemArea code into a separate crate
